+++ package.json
@@ -10,8 +10,10 @@
     "cra-template": "1.0.3",
     "firebase": "^7.24.0",
     "history": "^4.10.0",
+    "rc-progress": "^3.1.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-dropzone": "^11.3.2",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
