+++ src/components/UploadComponent/UploadComponent.css
@@ -1,0 +1,60 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    font-family: sans-serif;
+}
+  
+.container > p {
+    font-size: 1rem;
+}
+  
+.container > em {
+    font-size: .8rem;
+}
+  
+.dropzone {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    border-width: 2px;
+    border-radius: 2px;
+    border-color: #eeeeee;
+    border-style: dashed;
+    background-color: #fafafa;
+    color: #bdbdbd;
+    outline: none;
+    transition: border .24s ease-in-out;
+}
+  
+.dropzone:focus {
+    border-color: #2196f3;
+}
+  
+.dropzone.disabled {
+    opacity: 0.6;
+}
+  
+.selected_document_list {
+    margin-left: 5px;
+    cursor: pointer;
+}
+
+.dropzone-drag-option {
+    font-size: 30px;
+    margin-bottom: 10px;
+    margin-top: 10px;
+}
+
+.dropzone-select-option {
+    text-decoration: underline;
+    margin-top: 10px;
+    margin-bottom: 10px;
+}
+
+.upload-button-container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: flex-end;
+}
