+++ src/views/CasePage/CasePage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Sidebar from "../../components/Sidebar/Sidebar";
-import UploadButton from "../../components/UploadButton/UploadButton";
+import UploadComponent from "../../components/UploadComponent/UploadComponent";
 import formatDate from "../../assets/helpers/formatDate";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faChevronLeft } from "@fortawesome/free-solid-svg-icons";
@@ -88,7 +88,7 @@ export default class CasePage extends React.Component {
                     </tr>
                   </thead>
                   <tbody>
-                    {this.state.documents.map((onboard, i) => (
+                    {this.state.case.documents.map((onboard, i) => (  //Replaced documents -> case.documents, TODO: Check the difference between them
                       <React.Fragment
                         key={onboard.first_name + onboard.last_name + " " + i}
                       >
@@ -127,66 +127,8 @@ export default class CasePage extends React.Component {
                             borderTop: "none",
                           }}
                         >
-                          <td colSpan={5}>
-                            <img
-                              id={`frame-${i}`}
-                              title={onboard.name}
-                              alt="Original Document"
-                              src={onboard.file_link}
-                            ></img>
-                            <p>
-                              <b>Uploaded Document</b>
-                            </p>
-                            {onboard.translated_document_link ? (
-                              <div>
-                                <p>
-                                  <a
-                                    href={onboard.translated_document_link}
-                                    target="_blank"
-                                    rel="noopener noreferrer"
-                                  >
-                                    {onboard.name}
-                                  </a>
-                                </p>
-                                <img
-                                  id={`frame-translated-${i}`}
-                                  title={onboard.name}
-                                  alt="Translated Document"
-                                  src={onboard.translated_document_link}
-                                ></img>
-                              </div>
-                            ) : (
-                              <p>-</p>
-                            )}
-                            <br></br>
-                            <div className="uk-clearfix">
-                              <div className="uk-float-right">
-                                {onboard.translated_document_link ? (
-                                  <>
-                                    <button
-                                      className="uk-button uk-button-primary"
-                                      style={{ marginRight: "5px" }}
-                                    >
-                                      View Certification
-                                    </button>
-                                    <button className="uk-button uk-button-primary">
-                                      Undo Complete
-                                    </button>
-                                  </>
-                                ) : (
-                                  <>
-                                    <button
-                                      className="uk-button uk-button-primary"
-                                      style={{ marginRight: "5px" }}
-                                    >
-                                      Complete Task
-                                    </button>
-                                    <UploadButton document={onboard} />
-                                  </>
-                                )}
-                              </div>
-                            </div>
-                          </td>
+                          {/* Return <td> component containing upload section */}
+                          <UploadComponent document={onboard} index={i}/>
                         </tr>
                       </React.Fragment>
                     ))}
